+++ packages/page-navigation/src/GoToFirstPageMenuItem.tsx
@@ -9,18 +9,15 @@
 import * as React from 'react';
 import { LocalizationContext, MenuItem } from '@react-pdf-viewer/core';
 
+import { RenderGoToPageProps } from './types/index';
 import UpArrowIcon from './UpArrowIcon';
 
-export interface GoToFirstPageMenuItemProps {
-    onClick(): void;
-}
-
-const GoToFirstPageMenuItem: React.FC<GoToFirstPageMenuItemProps> = ({ onClick }) => {
+const GoToFirstPageMenuItem: React.FC<RenderGoToPageProps> = ({ isDisabled, onClick }) => {
     const l10n = React.useContext(LocalizationContext);
     const label = l10n && l10n.pageNavigation ? l10n.pageNavigation.goToFirstPage : 'First page';
 
     return (
-        <MenuItem icon={<UpArrowIcon />} onClick={onClick}>
+        <MenuItem icon={<UpArrowIcon />} isDisabled={isDisabled} onClick={onClick}>
             {label}
         </MenuItem>
     );