+++ packages/page-navigation/src/GoToLastPageButton.tsx
@@ -10,18 +10,18 @@ import * as React from 'react';
 import { Button, LocalizationContext, Position, Tooltip } from '@react-pdf-viewer/core';
 
 import DownArrowIcon from './DownArrowIcon';
-import { RenderGoToLastPageProps } from './GoToLastPage';
+import { RenderGoToPageProps } from './types/index';
 
 const TOOLTIP_OFFSET = { left: 0, top: 8 };
 
-const GoToLastPageButton: React.FC<RenderGoToLastPageProps> = ({ onClick }) => {
+const GoToLastPageButton: React.FC<RenderGoToPageProps> = ({ isDisabled, onClick }) => {
     const l10n = React.useContext(LocalizationContext);
     const label = l10n && l10n.pageNavigation ? l10n.pageNavigation.goToLastPage : 'Last page';
 
     return (
         <Tooltip
             position={Position.BottomCenter}
-            target={<Button onClick={onClick}><DownArrowIcon /></Button>}
+            target={<Button isDisabled={isDisabled} onClick={onClick}><DownArrowIcon /></Button>}
             content={() => label}
             offset={TOOLTIP_OFFSET}
         />