+++ packages/page-navigation/src/GoToNextPageButton.tsx
@@ -9,19 +9,19 @@
 import * as React from 'react';
 import { Button, LocalizationContext, Position, Tooltip } from '@react-pdf-viewer/core';
 
-import { RenderGoToNextPageProps } from './GoToNextPage';
 import NextIcon from './NextIcon';
+import { RenderGoToPageProps } from './types/index';
 
 const TOOLTIP_OFFSET = { left: 0, top: 8 };
 
-const GoToNextPageButton: React.FC<RenderGoToNextPageProps> = ({ onClick }) => {
+const GoToNextPageButton: React.FC<RenderGoToPageProps> = ({ isDisabled, onClick }) => {
     const l10n = React.useContext(LocalizationContext);
     const label = l10n && l10n.pageNavigation ? l10n.pageNavigation.goToNextPage : 'Next page';
 
     return (
         <Tooltip
             position={Position.BottomCenter}
-            target={<Button onClick={onClick}><NextIcon /></Button>}
+            target={<Button isDisabled={isDisabled} onClick={onClick}><NextIcon /></Button>}
             content={() => label}
             offset={TOOLTIP_OFFSET}
         />